fix: minor changes, move page metadata to own file
