+++ db/index.ts
@@ -10,3 +10,5 @@ import * as schema from "./schema";
 const queryClient = postgres(env.DATABASE_URL);
 
 export const db = drizzle(queryClient, { schema });
+
+export default db;
